Move pingStory to story page
